function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-servicebus==0.50.2
psycopg2-binary
sendgrid==6.3.0
pylint==2.5.2
python-dateutil==2.8.1
python-http-client==3.2.7